Fixed bug where AI terminates early
Bug existed where AI terminates even though there are valid moves, because utility value for the losing move is higher than the moves that allow the AI to continue. Because now the bestMove will return null when the game ends, a check is done at PlayerSkeleton for the null value. Fix results in a rather significant jump in number of rows cleared.